Change how fingerprint is arrived at in Swish component
